+++ src/LiveCharts.Core/DataSeries/PieSeries.cs
@@ -129,7 +129,7 @@ namespace LiveCharts.Core.DataSeries
                 Duration = AnimationsSpeed == TimeSpan.MaxValue ? chart.View.AnimationsSpeed : AnimationsSpeed,
                 AnimationLine = AnimationLine ?? chart.View.AnimationLine
             };
-            var originalDuration = timeLine.Duration.TotalMilliseconds;
+            var originalDuration = (float)timeLine.Duration.TotalMilliseconds;
             var originalAnimationLine = timeLine.AnimationLine;
             var i = 0;
 
@@ -165,7 +165,7 @@ namespace LiveCharts.Core.DataSeries
                 {
                     timeLine = AnimationExtensions.Delay(
                         // ReSharper disable once PossibleMultipleEnumeration
-                        originalDuration, originalAnimationLine, i / (double)PointsCount, DelayRule);
+                        originalDuration, originalAnimationLine, i / (float)PointsCount, DelayRule);
                 }
 
                 current.Coordinate.TotalStack = stack;
